+++ src/ImportCss/ImportCss.mjs
@@ -1,11 +1,13 @@
 /**
  * @interface
+ * @deprecated
  */
 export class ImportCss {
     /**
      * @param {string} url
      * @returns {Promise<CSSStyleSheet>}
      * @abstract
+     * @deprecated
      */
     import(url) { }
 }